change design pages

+++ vite-project/src/components/ManageCategories.jsx
@@ -6,11 +6,13 @@ import {
   deleteCategory,
 } from "../services/api";
 import toast, { Toaster } from "react-hot-toast";
+import { FaEdit, FaTrash, FaPlus } from "react-icons/fa";
 
 const ManageCategories = () => {
   const [categories, setCategories] = useState([]);
   const [newCategory, setNewCategory] = useState("");
   const [editingCategory, setEditingCategory] = useState(null);
+  const [showModal, setShowModal] = useState(false);
 
   useEffect(() => {
     fetchCategories();
@@ -25,45 +27,42 @@ const ManageCategories = () => {
     }
   };
 
-  const handleAddCategory = async () => {
+  const resetForm = () => {
+    setNewCategory("");
+    setEditingCategory(null);
+  };
+
+  const handleAddOrUpdateCategory = async () => {
+    if (!newCategory) {
+      toast.error("Category title cannot be empty");
+      return;
+    }
     try {
-      if (!newCategory) {
-        toast.error("Category title cannot be empty");
-        return;
+      if (editingCategory) {
+        await updateCategory(editingCategory._id, { title: newCategory });
+        toast.success("Category updated successfully!");
+      } else {
+        await addCategory({ title: newCategory });
+        toast.success("Category added successfully!");
       }
-      await addCategory({ title: newCategory });
-      setNewCategory("");
+      resetForm();
+      setShowModal(false);
       fetchCategories();
-      toast.success("Category added successfully!");
     } catch (error) {
-      console.error("Add Category Error: ", error);
-      toast.error("Failed to add category");
+      toast.error(`Failed to ${editingCategory ? "update" : "add"} category`);
     }
   };
 
   const handleEditCategory = (category) => {
     setEditingCategory(category);
     setNewCategory(category.title);
-  };
-
-  const handleUpdateCategory = async () => {
-    try {
-      if (!newCategory) {
-        toast.error("Category title cannot be empty");
-        return;
-      }
-      await updateCategory(editingCategory._id, { title: newCategory });
-      setEditingCategory(null);
-      setNewCategory("");
-      fetchCategories();
-      toast.success("Category updated successfully!");
-    } catch (error) {
-      console.error("Update Category Error: ", error);
-      toast.error("Failed to update category");
-    }
+    setShowModal(true);
   };
 
   const handleDeleteCategory = async (id) => {
+    const confirmed = window.confirm("Are you sure you want to delete this category?");
+    if (!confirmed) return;
+
     try {
       await deleteCategory(id);
       fetchCategories();
@@ -74,75 +73,101 @@ const ManageCategories = () => {
   };
 
   return (
-    <div className="container mx-auto my-8 p-6 bg-white shadow-lg rounded-lg ">
+    <div className="container mx-auto my-8 p-4 lg:p-8 flex flex-col items-center shadow-lg rounded-lg">
       <Toaster />
-      <h2 className="text-2xl md:text-3xl font-bold mb-6 text-gray-700">
-        Manage Categories
-      </h2>
-      <div className="flex flex-col md:flex-row mb-4">
-        <input
-          type="text"
-          value={newCategory}
-          onChange={(e) => setNewCategory(e.target.value)}
-          placeholder="Category Title"
-          className="flex-grow p-2 border border-gray-300 rounded-md md:rounded-l-md focus:outline-none focus:ring focus:ring-indigo-200 mb-2 md:mb-0 md:mr-2"
-        />
-        <button
-          onClick={editingCategory ? handleUpdateCategory : handleAddCategory}
-          className={`px-4 py-2 text-white rounded-md md:rounded-r-md ${
-            editingCategory
-              ? "bg-yellow-500 hover:bg-yellow-600"
-              : "bg-indigo-600 hover:bg-indigo-700"
-          }`}
-        >
-          {editingCategory ? "Update Category" : "Add Category"}
-        </button>
-      </div>
-
-      <div className="overflow-x-auto">
-        <table className="min-w-full bg-white border border-gray-300 rounded-md">
+      <h2 className="text-3xl text-amber-900  font-bold mb-12 ">Manage Categories</h2>
+      <button
+        onClick={() => setShowModal(true)}
+        className="fixed bg-transparent text-xl hover:underline font-semibold right-14 border-[#853e3e] text-[#936767] mb-5 px-3 py-10"
+      >
+        Add Category
+      </button>
+      <div className="overflow-x-auto w-full">
+        <table className="table-auto border w-full text-center shadow-md rounded-lg mb-8">
           <thead>
-            <tr className="bg-gray-100 text-gray-600 text-left">
-              <th className="py-3 px-4 uppercase font-semibold text-sm">
-                Category Title
-              </th>
-              <th className="py-3 px-4 uppercase font-semibold text-sm text-right">
-                Actions
-              </th>
+            <tr className="bg-[#e2d6d6] tab-border-2 text-sm">
+              <th className="px-4 py-2 border">ID</th>
+              <th className="px-4 py-2 border">Category Title</th>
+              <th className="px-4 py-2 border">Actions</th>
             </tr>
           </thead>
           <tbody>
-            {categories.map((category) => (
-              <tr key={category._id} className="border-b hover:bg-gray-50">
-                <td className="py-3 px-4">{category.title}</td>
-                <td className="py-3 px-4 text-right">
-                  <div className="flex flex-col sm:flex-row justify-end space-y-2 sm:space-y-0 sm:space-x-2">
-                    <button
-                      onClick={() => handleEditCategory(category)}
-                      className="bg-blue-500 hover:bg-blue-600 text-white px-3 py-1 rounded-md"
-                    >
-                      Edit
-                    </button>
-                    <button
-                      onClick={() => handleDeleteCategory(category._id)}
-                      className="bg-red-500 hover:bg-red-600 text-white px-3 py-1 rounded-md"
-                    >
-                      Delete
-                    </button>
-                  </div>
-                </td>
-              </tr>
-            ))}
-            {categories.length === 0 && (
+            {categories.length > 0 ? (
+              categories.map((category, index) => (
+                <tr key={category._id} className="border-t hover:bg-gray-100">
+                  <td className="px-4 py-2 border">{index + 1}</td>
+                  <td className="px-4 text-xl py-2 border">{category.title}</td>
+                  <td className="px-4 py-2 border">
+                    <div className="flex justify-center">
+                      <button
+                        className="text-[#612121] p-2 flex items-center justify-center"
+                        onClick={() => handleEditCategory(category)}
+                        title="Edit Category"
+                      >
+                        <FaEdit size={20} />
+                      </button>
+                      <button
+                        className="text-[#a42323] p-2 flex items-center justify-center"
+                        onClick={() => handleDeleteCategory(category._id)}
+                        title="Delete Category"
+                      >
+                        <FaTrash size={16} />
+                      </button>
+                    </div>
+                  </td>
+                </tr>
+              ))
+            ) : (
               <tr>
-                <td colSpan="2" className="text-center py-4 text-gray-500">
-                  No categories found.
+                <td colSpan="3" className="text-center py-4">
+                  No categories found
                 </td>
               </tr>
             )}
           </tbody>
         </table>
       </div>
+
+      {showModal && (
+        <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-90">
+          <div className="modal-box bg-amber-900 p-3 rounded max-w-2xl w-full mx-4">
+          <div className="bg-white shadow-lg rounded-lg p-8 max-w-4xl w-full">
+
+            <h3 className="text-2xl  text-[#612121] text-center font-bold mb-4">
+              {editingCategory ? "Edit Category" : "Add Category"}
+            </h3>
+            <input
+  type="text"
+  value={newCategory}
+  onChange={(e) => setNewCategory(e.target.value)} 
+  placeholder="Category Title"
+  className="input input-bordered w-full mb-4 border-blue-900 focus:border-amber-900 focus:ring focus:ring-amber-900 focus:ring-opacity-60"
+/>
+
+            <div className="flex justify-center gap-2  mt-6">
+            <button
+                onClick={() => setShowModal(false)}
+                className="btn  border-amber-900 bg-transparent  hover:bg-amber-700  text-blue-950 font-bold py-2 px-2 rounded-lg w-20"
+                >
+                Cancel
+              </button>
+          
+              <button
+                onClick={handleAddOrUpdateCategory}
+                className="btn bg-amber-900 hover:bg-amber-700  w-20  text-white font-bold py-2 px-2 rounded-lg"
+              >
+                {editingCategory ? "Update " : "Add "}
+              </button>
+         
+            </div>
+
+
+            </div>
+
+          </div>
+
+        </div>
+      )}
     </div>
   );
 };
